requirements.txt
beautifulsoup4>=4.12.3
lxml>=5.2.2
requests>=2.32.3
tqdm>=4.66.4

hf_xet

fastapi>=0.111.0
uvicorn[standard]>=0.30.1

SQLAlchemy>=2.0.31
pydantic>=2.7.4
python-dotenv>=1.0.1

spacy>=3.7.5
transformers>=4.43.3
torch>=2.3.1

pandas>=2.2.2
matplotlib>=3.9.0
typer>=0.12.3